API/requirements.txt
Django==3.1.6 
djangorestframework==3.12.2
django-cors-headers==3.2.0
django-ckeditor==6.2.0
python-dotenv==0.17.1
PyJWT==2.1.0
gunicorn==20.1.0